+++ dart/lib/src/sentry.dart
@@ -94,6 +94,9 @@ class Sentry {
   /// Check if the current Hub is enabled/active.
   static bool get isEnabled => currentHub.isEnabled;
 
+  /// Last event id recorded by the current Hub
+  static SentryId get lastEventId => currentHub.lastEventId;
+
   /// Adds a breacrumb to the current Scope
   static void addBreadcrumb(Breadcrumb crumb, {dynamic hint}) {
     currentHub.addBreadcrumb(crumb, hint: hint);