+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Ref: added Transport #123
 - Feat: apply sample rate
 - Ref: execute before send callback
+- Feat: add lastEventId to the Sentry static API
 - Feat: addBreadcrumb on Static API
 
 # `package:sentry` changelog